HiSilicon V5 + aarch64 support (3519DV500 / 3516CV610) (#168)
* chipid: enable HiSilicon V5 dispatch on aarch64

The V5 family (Hi3516CV610 / Hi3516CV613 / Hi3516DV500 / Hi3519DV500)
all share UART0 @ 0x11040000 and SC_CTRL @ 0x11020000. The case was
previously gated by `#ifdef __arm__` together with the legacy V1..V4
dispatches, which made the 64-bit V5 chips (Hi3516DV500, Hi3519DV500)
fall through to generic_detect_cpu() on aarch64.

Lift the V5 case to the top of hw_detect_system()'s switch so it's
reachable from both ARMv7 (Hi3516CV610/613) and aarch64
(Hi3516DV500/3519DV500). Legacy V1..V4 cases stay 32-bit-only.

Verified on Hi3519DV500 (aarch64): SCSYSID0 at 0x11020EE0 reads
0x3519D500, which hal_hisi.c's existing chip table already maps to
"3519DV500" / HISI_OT.


* backup, tools: fix 64-bit and null-deref bugs surfaced by aarch64 build

Three pre-existing bugs invisible on 32-bit ARM / musl but fatal on
aarch64 / glibc:

backup.c: `uint32_t payload` is passed by address to fread_to_buf()
which takes `size_t *`. ARMv7 size_t == uint32_t so the mismatch never
fired; on aarch64 size_t == uint64_t and the build fails outright.
Switch the local to size_t.

tools.c (get_god_pid): `strncpy(shortname, maxname, shortsz)` is called
unconditionally, but the only in-tree caller (get_god_app) passes
(NULL, 0). aarch64 glibc faults on strncpy with NULL dst even when
n == 0. Guard with `if (shortname && shortsz)`.

tools.c (get_pid_cmdline): after fopen("/proc/PID/cmdline") fails
(common when a PID exits mid-/proc walk), the function falls through to
fclose(fp) on fp == NULL. Restructure to early-return on NULL fp.


* mtd: stream-hash UBI volumes to avoid OOM on large rootfs

cb_mtd_info() computed each UBI volume's SHA1 by calling
read_ubi_volume() with a single malloc(data_bytes), then SHA1'ing the
buffer in one shot. On targets whose rootfs UBI volume approaches
system RAM size the malloc OOMs the process: Hi3519DV500 has a
data_bytes = 0x1D008000 (~464 MB) rootfs vol on 512 MB total, which
makes a bare ipctool exit 137 immediately (kernel oom-killer log:
"Killed process N (ipctool) total-vm:495608kB, anon-rss:450564kB").

Add sha1_ubi_volume() that streams the volume in 64 KB chunks through
SHA1Init / SHA1Update / SHA1Final, and switch cb_mtd_info() to it.

read_ubi_volume() is left untouched because backup.c legitimately
needs the full buffer to dump volume contents.


* clocks: add V5 family (3516CV610 / 3519DV500) — 4 PLLs + DDR + SSMOD + HPM

Decodes the full clock tree of the HiSilicon V5 (HISI_OT) family.
Same clocks_family_v5 covers both 32-bit (3516CV610/3516CV613) and
64-bit (3516DV500/3519DV500) variants since they share the same CRG /
PHY / HPM register map.

Wired through HISI_OT in clocks.c's families[] dispatch.

PLLs (CRG @ 0x11010000):
  APLL (CPU)       ctrl 0x000/0x004, lock 0x038
  VPLL (Video)     ctrl 0x080/0x084, lock 0x0B8
  DPLL (DDR)       ctrl 0x180/0x184, lock 0x1B8
  EPLL (Ethernet)  ctrl 0x200/0x204, lock 0x238
Same V4 field layout (FRACDIV[23:0], POSTDIV1[26:24], POSTDIV2[30:28],
FBDIV[11:0], REFDIV[17:12]); lock bit 4 = apll_lock_final per the
vendor's apll_lock_status union in bsp/components/gsl/boot/reset.c.

DDR data rate is a 3-stage chain (CRG ddr_cksel → DPLL → PHY
dficlk_ratio) that doesn't fit struct mux_info, so this also adds an
optional `extra(cJSON *root, bool brief)` callback to clock_family.
V5's v5_extra() decodes:
  - DDR data rate     = phy_clk_mhz * (1 << dficlk_ratio)
                        with ddr_cksel @ CRG+0x2000 [18:16],
                             dficlk_ratio @ DDR_PHY0+0x78 [1:0]
  - DRAM type         @ DDR_PHY0+0x2C [3:0]
  - SSMOD (DPLL SSC)  @ CRG+0x190 (cken bit 0, disable bit 2)
  - per-site HPM      @ HPM_BASE 0x1102B000 (NPU/MDA/CORE, 4 samples each)
Canonical HPM binning value @ SYSCTRL+0x340 [9:0] goes through the
standard hpm_info table.

Ground-truth on Hi3519DV500 demo board: APLL 750 MHz, DPLL 664 MHz,
DDR4-1333 (1328 MT/s), SSMOD active, HPM core_hpm_value=300 (high).


---------

+++ src/backup.c
@@ -1002,7 +1002,7 @@ static int do_upgrade(const char *filename, bool force) {
         goff = strtol(offset->valuestring, 0, 16);
         printf("Using offset: %#X\n", goff);
     }
-    uint32_t payload;
+    size_t payload;
 
     const cJSON *parts = cJSON_GetObjectItemCaseSensitive(json, "partitions");
     ASSERT_JSON(parts);
